+++ Assets/CHANGELOG.md
@@ -6,3 +6,10 @@
         - Sprinting
         - Rotating
         - Collision
+
+#Changelog 0.0.1-devstage
+    - Added Jump and Run
+    - Added Colors
+    - Edited Player
+        - Jumping (with Physics. Only jumpable if object is on "Ground" Layer)
+        - Crouching (Smooth, from AI)